Small grammar fixes in timers-7.md
I went through the first third of timers-7.md and made some small changes to correct grammar usage and try to make it flow better. Please let me know if any of my changes affect the meaning or style of the writing, as this was not my intent.

If these sorts of changes are desirable, I can go through the rest of the file for grammar (as well as other files).